+++ examples/with-angular-thirdpartyemailpassword/src/app/home/home.component.ts
@@ -1,6 +1,6 @@
 import { AfterViewInit, Component } from "@angular/core";
 
-import * as SuperTokens from "supertokens-website";
+import * as Session from "supertokens-web-js/recipe/session";
 
 @Component({
     selector: "app-home",
@@ -19,14 +19,14 @@ export class HomeComponent implements AfterViewInit {
     }
 
     async getUserInfo() {
-        this.session = await SuperTokens.doesSessionExist();
+        this.session = await Session.doesSessionExist();
         if (this.session) {
-            this.userId = await SuperTokens.getUserId();
+            this.userId = await Session.getUserId();
         }
     }
 
     async onLogout() {
-        await SuperTokens.signOut();
+        await Session.signOut();
         window.location.reload();
     }